+++ okra_py/income.py
@@ -1,4 +1,4 @@
-from .auth import Initializer
+from .base import Initializer
 
 class OkraIncome(Initializer):
     """
